TpProxy: don't forcibly fail method calls because of application-specific invalidation; also defer callback due to DBusGProxy destruction until main loop is re-entered.
This changes the behaviour of a closed channel or disconnected connection,
for instance. Previously, we'd force the method call to fail when we got the
Closed or StatusChanged signal, then ignore the real reply when it came in.

We now allow the reply to method calls that were made before invalidation
to be processed as usual, although method calls made after invalidation will
fail immediately.

When a DBusGProxy is destroyed (when name ownership is lost) we still force the
method call to fail (on the basis that the service has fallen off the bus, so
can't possibly send us a reply any more), but we now defer calling the callback
until an idle handler, for the usual reentrancy reasons.

20080111161922-53eee-7fe60d38438f4f75517514fd2e0ca4dd2a8f3ef0.gz