Require TpChannel to have a TpConnection, and expose it as a property.
While it's technically possible to use a TpChannel without a TpConnection
(inspect-channel currently does), it's unlikely to be a good idea in any
non-trivial cases - handle validity is closely tied to the connection's
lifetime, and if you don't discard handles promptly, you'll get confused.
See fd.o #14188 and (maybe) Gnome #511063.

20080124125025-53eee-57945d04215ef5ed17829019fd837b9ee6c5e8eb.gz