Make registering DBus.Properties for the mixin more declarative.
The interfaces now just have a gchar * for the name (this avoids having to
fill in computed values, the quark and GType, in class_init - everything
can be static initialization data) and the mixin walks the interface list
at runtime to tag the interface implementation with its corresponding
interface info.
Also, use different user_data for the getter and setter functions, so those
two functions don't have to be tightly coupled.

20080218122213-53eee-c73c34ee631e11c208e6b3020649001a1ba56c7f.gz