channel: delay creating priv->sessions until after GetSessionHandlers has returned.
This lets us use it as a guard against adding a session due to
NewMediaSessionHandler. Previously, we could get this:

s-e: GetSessionHandlers?
CM: NewSessionHandler! /foo/bar
CM: here are your session handlers: ["/foo/bar",]

and s-e would create two sessions for the same session-handler /foo/bar.
This caused s-e to call Ready() twice, which (due to a Gabble bug) caused
double emission of each NewMediaStreamHandler signal, resulting in s-e
thinking the CM was trying to give it two streams with the same ID and
cancelling the call.

20080222151358-53eee-8cd44c4a30e0cf1528743f1fc2f2243a05f86141.gz