TpTextMixin: drop the futile attempt to limit memory consumption
There are so many ways to induce a typical connection manager to consume memory
(opening large numbers of text channels from different contacts; sending large
numbers of invitations to different chatrooms; starting large numbers of VoIP
calls) that guarding against this one in particular, and dropping messages as
a result, doesn't seem worthwhile at all. In particular, arbitrary message
truncation is the wrong way to do it.

If we do anything about memory limiting, we should (a) do it right, and (b)
limit memory per connection rather than globally. If we really care about this,
the way forward would probably be to use a separate process per connection
(the Telepathy spec allows for this, although many clients probably get it
wrong since we've never implemented it before).

20080408143429-53eee-4b0a6694439aec7582e1be639a252c36d2fa8eae.gz