Avoid a misleading debug message every time a pending call completes, and probably fix detection of crashing services
Due to wrong use of != instead of == in a test, we would previously emit a
misleading debug message every time a pending call finished when a reply *had*
been queued, instead of emitting that message when it had *not* (due to a
redundant test elsewhere, this was mostly harmless).

Similarly, we would ignore pending calls "finishing" when a reply had not yet
been queued, which I suspect means we wouldn't detect a service crashing (or
other name-owner-lost events) as intended.

20080528085115-53eee-83c8d6fed72315f6b424933dd1f9d95082ab7e27.gz