Add tp_text_mixin_has_pending_messages
The proposal of "closed text channels are resurrected if they still have
incoming messages" is much easier to implement if we do this - it's possible
to use tp_text_mixin_list_pending_messages(), but that does wasteful copying.

The idea of returning the first sender handle is that callers can use it
as the initiator of the resurrected channel.

20080724172126-53eee-2562dd209968bb4508582101b89be3db37d178b9.gz