BaseConnection: set self_handle to 0 just after emitting status-changed(Disconnected)
Zeroing it in finish_shutdown() broke connection managers that unreffed
the connection before calling that method, as base_connection_dispose
asserts that self_handle == 0 (because the handle factories complain if
there are outstanding refs when they are destroyed).

Zeroing it in base_connection_dispose broke Gabble, because
gabble_connection_dispose asserts that the self_handle is already 0.
This is fd.o #17751, as dispose is conceptually the right place to drop
the ref.