TpConnection: add handle ref/hold/release functionality
This centralizes handle ref tracking, because otherwise calling
ReleaseHandles is inherently unsafe - you never know whether
another part of the process is still using the handles.

In this implementation the handles are stored in a data-slot on the
DBusConnection.

See mailing list thread "Uniqueness of objects", starting
with Message-ID: <20080930200125.GA15304@carbon.pseudorandom.co.uk>