TpContact: on success, don't give the callback a free ref to each contact
It seems saner to require the callback to take a ref itself if it wants
to keep the object (and in fact examples/client/inspect-contact.c failed
to unref it, so it's correct for the new semantics).

Also, on fatal errors, give the callback a bit more information about
what was going on, by telling it the handles or IDs that were requested
in the first place (via the "failed handles/IDs" arguments).