TpChannel: invalidate during introspect if something goes horribly wrong
GetChannelType, GetInterfaces and GetHandle are all fundamental, have
all been in the spec as long as I can remember, and all have no
documented error conditions. So, if one of these fails, we can declare
that the Channel is, in fact, not a Channel. This allows us to detect
(for instance) object paths that point into the void rather than to a
working channel.