Remove FIXME about tp_asv_get_boolean accepting ints
If we want to make it more tolerant in the future, we can do so, but it
doesn't seem necessary.  In the integer case, the tolerance is needed
for glib because it doesn't have 16-bit ints, and to make short Python
scripts easier.  The only language I can think of where treating ints as
bools would be useful is Perl, but since there are only two boolean
values forcing people to use magic DBus constants in that case isn't so
bad.