tp_connection_manager_end_introspection: always emit got-info
Either we succeeded (so we should tell the user), or we failed and
have no cached info (so we should tell the user to stop waiting),
or we failed but already had cached info (so there's no harm in
telling the user).

This appears to fix the fact that got-info was not always emitted.