Install the type infomation earlier
If the type information is installed last then it is possible for dbus-glib to
assert because the dbus-glib containers haven't been initialized yet.  This is
done by installing type info, getting a bus connection, or registering an
object, which isn't always done before object construction.

[Patch modified to install the information just after variable
declarations, rather than before as in Ross's original patch, in order
to be valid C89.]

Signed-off-by: Will Thompson <will.thompson@collabora.co.uk>