TpGroupMixin: allow channels to make the self-handle always removable
Removing the self-handle is (meant to be) the documented way to leave a
chatroom or call gracefully, and possibly with a message or reason code,
so we should support it in all cases. Sadly, that risks crashing
connection managers whose remove callback makes assertions about who
can be removed, so it has to be "opt-in".