fd.o #22182: fix a potential use-after-free in the callable example CM
In some environments (I'm not sure why my laptop isn't among them),
closing the stream causes the channel to remove the last ref to the
stream, resulting in a use-after-free and a possible segfault.

This could be avoided by holding another ref for the duration of
example_callable_media_stream_close, but Sjoerd and I agreed that this
simpler fix (re-ordering to put the possible free last) was more
appropriate.

Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>