Test that new_from_properties doesn't call GetAll
The artificially-constructed properties hash didn't have all of the
mandatory properties (Requested, Initiator* and TargetID were missing).
The test also used a service-side channel that didn't implement the
Properties interface. Finally, wrap the GObject property getter given to
the properties mixin with one that records that it's been called, and
assert that it hasn't been called once the channel becomes ready.