Make GetGroupFlags() failing abort introspection.
If we get as far as introspecting the Group interface of a non-existant
channel, or a channel which exists but was lying about having the Group
interface, introspection fails rather than assuming all the group
properties have empty values.

GetGroupFlags() has been around in its current form for a very long
time, so I think it's safe to assume that any channel which doesn't
implement it properly needs fixing.