Don't ping Group channels with GetInterfaces()
If we're about to introspect the Group interface anyway, there's no need
to call GetInterfaces() to check if the channel's alive: the Group
introspection will abort if the channel's dead or broken.

This saves an unnecessary roundtrip in the case of group channels for
which we have immutable properties; on the happy path, we can introspect
a group channel with just one call to GetAll("Chan.Iface.Group").