Mark default parameters as present
When creating a new connection BaseConnectionManager sets default parameters in
the parsed_params data structure, but didn't mark them as present. In practise
this meant that the Connection Managers didn't pass the default parameters on
the the Connection object they generate. Which in turns meant they rely on
having the knowledged about which parameters are default in two places (both in
the ConnectionManager object as a flag and in the Connection object as a
property that gets set by default), which seems unfortunate and makes adding
default values for parameters slightly fragile