When removing the last TpDBusDaemon from a DBusConnection, forcibly free the GSList ** too
You might think (and indeed I did) that
dbus_connection_free_data_slot (&daemons_slot) would automatically clear
out that slot on every DBusConnection, calling the registered
free-function for the old contents.

This is, as it turns out, untrue; you have to do that yourself, first.

This turns out to crash git master (the MC 4 branch) of Empathy during
startup: it briefly creates a TpDBusDaemon, then deallocates it again.
The GSList ** is placed in slot 2 on the DBusConnection. We then
deallocate slot 2. When dbus-glib allocates a new slot number for its
DBusGProxyManager objects, it is also told to use slot 2; it looks in
slot 2, finds our GSList ** still there, interprets it as a
DBusGProxyManager *, and crashes with an assertion failure.

Reviewed-by: Will Thompson <will.thompson@collabora.co.uk>