Re-usable test functions.
I found that I was wanting to execute the same test functions one after
another. This can't be done asyncronously unless you pass a list of functions
around which can be called when events happen.

This framework allows me to test tp_account_prepare by first creating an
account manager, and then using it to create an account.

I also made some whitespace-only changes to conform to the style guide.