Don't let MC5 activate itself during AM tests.
This sets up dbus to use a session.conf which doesn't include the
standard services.
This is the same trick that gabble's twisted tests use.

I'm not really sure whether tests/dbus/dbus-1/session.conf.in is the
best place to put the new file. If anyone thinks there is a better
place for it, shout.