Don't rely on enum types being unsigned
Not all compilers use unsigned types for enums even if all the values are
postive, as such an enum with 3 items doesn't always fit in a 2 bit field.
This can cause interesting bugs.

Force the type to be an unsigned int, this hopefully doesn't break
anything and fixes the bug without breaking ABI etc.

Reviewed-By: Simon McVittie <simon.mcvittie@collabora.co.uk>