Starting using the Index logstore to keep track of the logged messages
Now when pending messages via the Messages interface are examined,
the ones that are present in the Index wont' be sent to the log-manager again.

In the same way, on received signal, a message being in the Index won't be logged.

It also does some stale id clean-up.

DEBUG message have been improved.