Instead of guarding against callbacks and pointers being different sizes, just compile-time-assert it
In practice all sorts of code will fail if this compile-time assertion
fails. Also, the union we previously used is picked up by gtk-doc 1.14,
which complains that we haven't documented it... rather than trying to
work around that, we can just get rid of it, and reinstate it in the
unlikely event that we later care about a platform where callbacks are
not the same size as pointers.