tp_channel_get_identifier: always return non-NULL
Despite the documentation, TpChannel:identifier was not consistently NULL
for anonymous channels: for those with a TargetID D-Bus property (i.e.
all current, correct connection managers), it would be the empty string.

It seems simplest to guarantee it to be non-NULL, making unprepared
channels, and channels from old CMs, consistent with prepared channels
from new CMs (the common case).