Remove DBusGLib-1.0.gir requirement
No one has DBusGLib-1.0.gir, it doesn't ship. It's not at all crucial for
tp-glib, you can do everything you need to do with a TpDBusDaemon. This patch
removes the requirement for it and (skip)s any functions that pass dbus-glib
types as parameters or returns.

This currently breaks the build due to
https://bugzilla.gnome.org/show_bug.cgi?id=616375 -- which makes it impossible
to skip a GObject property, and thus one mention of DBusGConnection appears in
the GIR. You can remove this by hand and rerun 'make' and things will be fine.