fd.o #27780: TpAccount: when Connection may have changed, emit notify::connection
TpAccount doesn't automatically create a new TpConnection when the
connection_object_path member changes, only when someone actually asks
for it.

This has the surprising effect that notify::connection isn't
emitted when the object path is first discovered.

Note that this may result in too many signal emissions: we'll emit
notify::connection even if the result of getting the connection property
hasn't, in fact, changed. However, it's the simplest possible fix.