fd.o #23369: TpChannel: if the CM gives us a recognized error name, use it for invalidation
More specifically: if "error" is present in MembersChangedDetailed, *and* it
can be mapped to a GError domain and code, use that domain and code as
the invalidation reason. If it's absent or would be mapped to
TP_DBUS_ERROR_REMOTE_ERROR, continue to use TP_ERRORS_REMOVED_FROM_GROUP
as the invalidation reason.