fd.o #23369: TpConnection: don't use ConnectionError for invalidation if not understood
If ConnectionError can be mapped to a GError domain and code, use that domain
and code as the invalidation reason as before.

However, if it would be mapped to TP_DBUS_ERROR_REMOTE_ERROR,
we now choose the error for TpProxy::invalidated based on the
Connection_Status_Reason, as if ConnectionError had not been emitted -
this gives better invalidation reasons for connections that emit an
implementation-specific ConnectionError that we don't actually
understand.