fd.o #28043: explicitly link tests/* against dbus-glib
This fixes compilation with LDFLAGS=-Wl,--no-add-needed, with GNU gold
(where this is the default), and hopefully with recent Fedora (where this
appears to be the default for ld too).

When the last round of Debian "fails to build with GNU gold" reports came
in, nothing in tests/ used libdbus or dbus-glib. However, the test for
gnio-utils needs the dbus-glib specialized type system, even though it
doesn't actually connect to D-Bus, so it needs to be linked accordingly.

Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>