Don't announce new channels as a side-effect of creating the object
When creating a list or group channel because the subclass told us it
was created (or something was renamed to it), announce it explicitly
if appropriate.

When creating a list or group channel in set_list_received, we don't need
to announce it, because it'll be announced in a moment anyway.

When creating a list or group channel for a request, we don't want to
announce it. It'll be announced when appropriate - when the group is
actually created (for groups), or when the contact list is received
(for lists).