fd.o #29174: copy the 'echo' example CM to tests/lib
This example is no longer exemplary, because we want all Text implementors
to implement Messages, and we want all ConnectionManager implementors to
implement Protocol.

The regression tests use it as an example of a CM that has neither
feature, to make sure we're backwards-compatible, so we should keep it
around for the tests, but not promote it as a good example.