TpBaseChannel: add initiator arg to tp_base_channel_reopened()
When a channel is re-spawned, we need to switch its initiator to the contact
that caused the channel to be re-spawned. Rather than requiring all client
code to explicitly change the initiator property separately (which is
currently a construct-only param), just do it all in one function call.  This
makes the client code more straightforward and side-steps the complication
that the property setter can not ref the initiator handle while in
construction and defers this work to the 'constructed' vfunc.