TpBaseChannel: remove per-instance interfaces variable
In GabbleBaseChannel, there were two ways to set custom interfaces on a
channel: A) subclass the channel and set klass->interfaces, and B) set the
per-instance self->interfaces variable.  In the constructor, if the instance
member was not set, it copied the class variable to the instance member.
Because TpBaseChannel is based on GabbleBaseChannel, we inherited this
feature, but it turns out that this feature was never actually used, and the
only benefit it gives you is to allow you to create channels with different
interfaces without actually subclassing the channel.  This doesn't seem like a
compelling feature, and nothing in Gabble was actually using it, so I'm just
going to remove it.  For TpBaseChannel, he only supported way to set custom
interfaces will be to subclass and set klass->interfaces to the desired value.