Fix some references in TpBaseChannel documentation
The fields of TpBaseChannelClass being referred to are not properties,
so we have to use the '.' notation (for struct fields) for gtk-doc to
crosslink them properly.

The comment about TpBaseChannel dealing with freeing object_path is
obsolete now.

Referring to the 'close' vfunc as close() in its documentation confuses
gtk-doc, so let's reword it to not need that.