Add tp_base_channel_is_destroyed()
Quite a few channel implementations have clean-up that they need to
ensure happens exactly once, and so used to keep their own track of
whether they've been closed or not. This accessor allows them to
delegate tracking that to TpBaseChannel without needing to use
g_object_get (..., "channel-destroyed", ...);

Technically this accessor should live on TpExportableChannel, since
that's where the property is defined. But its semantics are easier to
explain if we can talk about tp_base_channel_destroyed(), so I put it
here.