BaseChannel: ref ourself in _destroyed() and _reopened()
Both of these functions emit the 'closed' signal. A reasonably sensible
thing for a handler for that signal to do is to unref the channel
object. Thus: these functions need to ensure that the channel stays
alive until they're finished messing with it.

(Actually, currently the latter doesn't touch the channel after emitting
'closed', but there's no harm in keeping these functions symmetrical.)