Add tp_base_channel_close()
This calls the ->close() vfunc, having checked that the channel is not
already destroyed. I found that a whole bunch of channels had their own
close function which is equivalent to Close() but checks is_destroyed()
first, so this seems like sensible boilerplate to move up to
TpBaseChannel.

Of course, some channels might need their own flag to say "close is in
progress" so that two overlapping calls to Close() don't break them, but
that's okay, they can still do that.