BaseProtocol: DEBUG sanitized parameter values
In the Gabble test suite, there's a neat hack where the name of the test
being run is passed to Gabble in the 'account' parameter, but ignored
because the 'resource' is also set. This makes it easy to search for the
section of the debug output relating to a particular test.

The move to BaseProtocol meant the parameter values were no longer
logged. This patch adds them to a relevant debug message.

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>