Clean up connecting to text channel signals.
There's really no point in logging which signal we failed to connect to,
because the reason will always be the same: the object is invalidated,
or the interface in question isn't implemented. Plus, there's no point
in trying to carry on if we'll ultimately bail out if any errors
occurred.

So, let's just bail out as soon as we encounter an error connecting to a
signal.