base-channel: don't unref the target handle in dispose if we didn't ref it
If you set TpBaseChannelClass->target_handle_type to something other
than NONE, then the target handle is reffed in constructed. This is
really useful. However, the handle is unreffed in dispose regardless
of handle type. So, we should make sure that we have a !NONE target
handle type before trying to unref it in dispose.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>