fd.o #31377: don't use tp_tests_proxy_run_until_dbus_queue_processed with multiple D-Bus connections
tp_tests_proxy_run_until_dbus_queue_processed tells us that our message
has got through dbus-glib, to the dbus-daemon and back. However, in the
presence of another parallel D-Bus connection, that doesn't guarantee
that dbus-daemon has polled the other connection's socket; in
particular, it might not have noticed its disconnection and sent us
NameOwnerChanged.

Now that we have an arbitrary timeout on every test, it's more reasonable to
just spin the main loop until the expected event happens; if it doesn't,
we'll just be killed by the timeout, with a useful stack trace.