fd.o #31631: fix confusion between TpClientChannelFactory and TpClientChannelFactoryInterface
The former (which didn't previously have a typedef) is a GObject which
implements TP_TYPE_CLIENT_CHANNEL_FACTORY. The latter is the vtable
struct for that interface (i.e. it's a class-like structure).

This is an API break (for anyone who was using channel factories, which
are relatively recent), but not an ABI break.

Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>