Ensure CodecsUpdated is always called when needed
For a CM calling SupportedCodecs after local codecs have been set is only
expected to contain the intersection of the remote codecs with the existing set
of codecs. Specifically it's unexpected for codecs to change their parameters.

This means there is a race when remote codecs are updated and at the same time
the local codecs are updated, as telepathy-farsight tried to ensure to only
send a set of codecs once (via either SupportedCodecs, CodecsUpdated or Ready).
To fix this always call CodecsUpdated when the list of codecs has changed after
the initial set of codecs was sent.