_tp_signalled_message_new: ignore sender in @parts, trust @sender instead
This is necessary in some pathological cases, for instance if we get a
message-sender and a message-sender-id, do a round-trip to get a TpContact
based on the ID, and discover that their handle does not match the
message-sender we started with (which could happen if the handle went
out of scope, on older CMs).

Also use tp_message_delete_key instead of NIHing it, remove an unnecessary
cast, and check that if sender is non-NULL, it is valid.