Simplify dispatching NameOwnerChanged signals.
Previously there was this weird model where TpDBusDaemon had a dict
mapping bus names being watched to a watch struct. The watch struct had
a callback and some user data. When the library user first added a watch
for a name, their callback would be shoved into a new struct that was
put into the dict, and all was bright. But when a second call to
_watch_name_owner() came along … a GArray of "sub watches" would be
created, the old callback copied into it and the new callback added to
it, and then the original struct's callback would be replaced by an
callback that iterates the subwatches.

I couldn't see any good reason not to always have a list of callbacks,
so deleted that code. Now all watches have a list of callbacks, and it's
a bit clearer.

Making this change found a nasty bug where, if one callback removes its
own watch, subsequent callbacks may not be called for this signal. Test
case and fix to follow.