TpSimplePasswordChannel: Add support for MaySaveResponse property
Add support to the simple password for the MaySaveResponse property that was
recently added to the SASLAuthentication spec.  This property defaults to
TRUE, and in the normal case where the API user doesn't deal directly with the
channel, it should do the right thing.  For applications that want different
behavior, they can subclass TpSimplepasswordChannel and set the property as
appropriate.