TpSimplePasswordManager: add prompt_for_channel_finish()
Instead of having both prompt_async() and prompt_for_channel_async() call the
same _finish() function, provide a separate function for the custom channel
case.  This function will also return the custom channel that was originally
passed to prompt_for_channel_async().  This way, the API user doesn't need to
carry around the channel in user_data in the case that they need to query the
channel for information after the prompt has completed.