Move away from id/type pair in favor of TplEntity
In the first rework step, TplEventSearchType was introduced with wrong
assomption. This item was merging entity type and event type together. To
make thing simple, this patch replace the pair with a TplEntity object that
contain all the required information about the Entity. This enable for more
complete logs in the future and fix bugs where get_entities() won't return
chatrooms.