BaseContactList: emit BlockedContactsChanged
Just like we only emit the ContactList/ContactGroups signals if the
connection the BaseContactList is attached to implements those
interfaces, so here we only emit BlockedContactsChange if
ContactBlocking is implemented. We do the hard work of building the hash
tables anyway. I don't think this is a big cost (and we have been
turning intsets into arrays and then immediately freeing them since this
code was written!).

The test case checks that blocking/unblocking individual contacts using
the old API causes this new signal to be emitted.