Fix crash in contacts_get_attributes with invalidated conn
If the connection was invalidated at an inopportune time, contacts_get_attributes could be called
with the connection already having been invalidated and dropped its interfaces. This would cause an
assert checking for support for the Conn.I.Contacts interface to be hit.

This is a simple but safe way to prevent that and related bugs. An invalidated connection never
needs to build contacts successfully.