Tests: include system $XDG_DATA_DIRS
The glib-networking proxy resolver uses GSettings. GSettings looks for
schemas in XDG_DATA_DIRS. GSettings gets really angry if a schema is
missing:

  GLib-GIO-ERROR **: Settings schema 'org.gnome.system.proxy' is not installed

Setting XDG_DATA_DIRS replaces the defaults, rather than adding to them.

While in principle it's nice to be isolated from the system, and while
it would be better to ensure we don't hit the network by having a dummy
proxy resolver, I'd like to be able to run the test suite.

Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>