tp_intset_iter_{init,reset}: demacroify
Previously, tp_intset_iter_init (resp. _reset) was a macro, which was
implemented using tp_intset_iter_init_inline, which was a static inline
function defined in the header.

This seems unnecessary, particularly for an iterator type documented to
be slow and old and semi-deprecated in favour of the faster variant,
which uses a regular function to init.