Annotate tp_errors_quark as deprecated.
The docstring says it's deprecated; it's been that way for years; the
macro that people actually use has long since expanded to
tp_error_quark.

This doesn't actually make gtk-doc shut up, as I hoped it would, because
it doesn't pay the blindest bit of notice to G_GNUC_DEPRECATED and
instead witters on about deprecation guards. Oh well.