Clean up tp_handle_set_foreach()
gtk-doc was complaining that it didn't recognise the (closure)
annotation. While it is documented to be correct on
<https://live.gnome.org/GObjectIntrospection/Annotations>, since the
function is (skip)ed anyway I don't see any particular reason to keep it
around. Besides, if we call the parameter 'user_data' then GI makes the
right assumption.

While looking at this I noticed that the implementation inexplicably
initializes its iterator closure twice.