configure: squash ‘AC_LANG_SOURCE call detected’ warning.
Since I upgraded my toolchain, autogen has started complaining:

  configure.ac:109: warning: AC_LANG_CONFTEST: no AC_LANG_SOURCE call detected in body
  ../../lib/autoconf/lang.m4:194: AC_LANG_CONFTEST is expanded from...
  ../../lib/autoconf/general.m4:2591: _AC_COMPILE_IFELSE is expanded from...
  ../../lib/autoconf/general.m4:2607: AC_COMPILE_IFELSE is expanded from...
  m4/tp-linker-flag.m4:18: TP_LINKER_FLAG is expanded from...
  m4/tp-linker-flag.m4:41: TP_ADD_LINKER_FLAG is expanded from...
  configure.ac:109: the top level

Reading
<http://www.flameeyes.eu/autotools-mythbuster/forwardporting/autoconf.html>
suggests that AC_LANG_PROGRAM() is implemented in terms of
AC_LANG_SOURCE() and so should work fine in this context, but it doesn't
seem to. In any case, since the program is empty we can just use the
single-argument form and squash the warning.