GroupMixin: correctly use contact-ids, not member-ids
For the past two-ish years telepathy-glib has included a key named
'member-ids' in the MembersChangedDetailed details dictionary. This is
wrong: the spec says it should be called contact-ids. Whoops.

I've checked Empathy, telepathy-logger, tp-glib and tp-qt4, and none of
them rely on the incorrect name. So … I think we're safe to just change
it.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=37116>
Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>