fixup! log_store_xml_get_events_for_file: replace edited messages.
* update to private _tpl_text_event_add_supersedes

Also addresses:
> I think you don't have to call g_hash_table_insert() here ? And same in next
> for loop.
Actually, this is for the chained supersedes use-case, but yeah: it's wrong.
Fixed it when I implemented event_queue_replace_and_supersede (),
and fixed the comment on superseded_links


> Use g_list_previous().

> Is that suppose to be funny ?
(apparently it wasn't funny; fixed)

> Your leaking a reference on the dumm_event. Maybe you want to rework you
> _add_supersedes() method into something like link_supersedes() that would take
> ownership ?

Created method event_queue_replace_and_supersede with similar feel
to event_queue_insert_sorted_after.
This is something that I wanted to do, but couldn't see a clean way to do.
Fixing the semantics of superseded_links made this easier.

> You don't check anymore if parsing event worked. A corrupted XML file will
> cause your code to crash.
>
*_parse_*() will currently only return NULL if g_object_new() does, but you're
right: if someone changes this behaviour later, I should be resilient to it.
Fixed.