TpTextChannel: fix preparing when message-sender-id is missing
6fe8293 introduced a regression where, if a CM does not include
message-sender-id in received messages, preparing a TpTextChannel for a
proxy which has pending messages will never complete.

As of that commit, the only reference to 'result' is owned by the call
to Get(..., 'PendingMessages'), which is released once
get_pending_messages_cb() returns. So the callback for
tp_connection_get_contacts_by_handle/tp_connection_get_contacts_by_id
never fires, because 'result' is specified as the weak object for the
call: the result dies immediately, and hence the get_contacts operation
is cancelled.

This patch makes the calls to get_contacts_by_* hold a reference to the
result which they pass as a weak object. This is an abuse of the
weak_object field to pass both an object and another pointer to the
callback. Arguably it would be clearer to use a context struct, and my
advocating the removal of exactly such a context struct triggered this
bug…

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=39172>