TextChannel: fail IncomingMessages if contacts can't prepare
If we can't prepare *any* contacts (as opposed to just some contacts
failing to prepare) then the channel is pretty broken. In the past, in
this case we would mark the IncomingMessages features as prepared
without failure, but silently drop the messages!

In practice this can only fail if the CM crashes or the connection goes
away behind our back. But it seems safer to mark the operation as
failed.

(Relatedly, if we just _complete() the GSimpleAsyncResult then TpProxy's
machinery marks the features as prepared or failed as appropriate.)