Do not try to use a NULL hash table
Normally, the contents are destroyed when the channel unrefs
them at start of its _dispose routine.
However, a content may be temporarily reffed while there are
async operations in progress (eg, got_content_media_properties),
When this happens, the unref does not destroy the content, which
will only be disposed of later, at which point the channel will
have been disposed, as the contents do not keep a ref on the
channel.

This patch fixes this by disposing the contents when the call
channel goes, regardless of whether a ref is still being held,
so that each content's conference will be put in the expected
order as well as other cleanup. Checks are added where neeed
to avoid using deleted data by the code that was still holding
a ref.

This scenario can happen rather easily when cancelling a call
just after requesting it in meego-app-im.