Tests: disable timeouts if we're being debugged
This is pretty evil, but I got sick of my gdb sessions being destroyed
because I forgot to set TP_TESTS_NO_TIMEOUT to prevent the SIGALRM going
off after 10 seconds or whatever. So this checks /proc/self/status to
see whether another process is tracing us. If a "TracerPid:" line is
present, and is followed by a non-zero pid, the timeout is disabled. If
the line is present but with a zero pid, if the line is missing, or if
the file is missing or reading it fails, we set the timeout as usual.