Contacts: avoid re-introspecting client types
My reading of the spec says that it's okay to leave /client-types out of
the result of GetContactAttributes().

My reading of the tp-glib documentation is that it's valid for
tp_contact_get_client_types() to return NULL even if
TP_CONTACT_FEATURE_CLIENT_TYPES is prepared. TpContact:client-types is
documented a little differently, so I'm changing it to match.