Contacts: don't wrongly mark SimplePresence prepared
Previously, if a broken CM included the SimplePresence/presence
attribute when TpContact didn't ask for it,
CONTACT_FEATURE_FLAG_PRESENCE would still be set, even if we weren't
listening to PresencesChanged. So this is broken.

This fix makes TpContact ignore unsolicited presence information in
GetContactAttributes replies. The other option was to bind to the signal
in that event, which seems … surprising.