Add tp_base_room_config_dup_channel()
This is handier than using the GObject property, and it transpires that
subclasses would like to use it.

It returns a ref rather than not because TpBaseRoomConfig only holds a
weak ref to the channel, so it would otherwise be quite easy for CM code
to accidentally try to use a dead pointer. (I fell into this trap while
working on Gabble.)