BaseRoomConfig: improve find_myself error checking
TpBaseRoomConfig stores a pointer to itself as qdata on its parent
channel. It did not previously NULL that pointer out when it was
disposed: it now does.

Correspondingly, I added an explicit check for NULL to find_myself().
While this is technically redundant with the TP_IS_BASE_ROOM_CONFIG()
check, I think it makes the error message clearer to distinguish between
the two.