gtk-doc: use no-tmpl flavour
This stops gtk-doc even generating the legacy .tmpl files, for a
marginally faster (and less crufty) documentation build process.

For some reason, this makes it start failing to parse the final argument
of functions annotated with _TP_GNUC_DEPRECATED_FOR(). Not overly
surprising that it can't parse them; the surprise is that no-tmpl
affects this. But since we already pass _TP_GNUC_DEPRECATED to
--ignore-decorators, it seems reasonable to also filter out
_TP_GNUC_DEPRECATED_FOR there.